test: merge microagent tests
- Keep legacy microagent test
- Add environment variable testing to new agents
- Add comprehensive tests for new microagent system
- Move all tests to tests/unit
